requirements.txt
attrs==26.1.0
cantera==3.2.0
CoolProp==7.2.0
coverage==7.13.5
dearpygui==2.1.1
execnet==2.1.2
gearbox==0.3.2
gekko==1.3.2
hupper==1.12.1
iapws==1.5.5
iniconfig==2.3.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.4
packaging==26.1
PasteDeploy==3.1.0
pluggy==1.6.0
pyfiglet==1.0.4
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-pretty==1.3.0
pytest-rich==0.2.0
pytest-xdist==3.8.0
rich==15.0.0
ruamel.yaml==0.19.1
scipy==1.17.1
setuptools==82.0.0
Tempita==0.6.0
typing_extensions==4.15.0
